:sparkles: 상세보기 테이블 수정

+++ src/main/java/com/uni/uni_erp/dto/sales/SalesSummaryDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class SalesSummaryDTO {
     private String itemName;
-    private int totalQuantity;
-    private double unitPrice; // Original unit price
+    private Integer totalQuantity;
+    private Integer unitPrice; // Original unit price
 
 }
